backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-openai==0.0.2
langgraph==0.0.20
openai==1.3.7
pandas==2.1.3
python-dotenv==1.0.0
sse-starlette==1.6.5

